Classpart fix

+++ src/main/java/seedu/taskmaster/model/record/StudentRecordListManager.java
@@ -81,7 +81,7 @@ public class StudentRecordListManager implements StudentRecordList {
      * Updates participation score of a student represented by their {@code nusnetId} to {@code score}.
      */
     @Override
-    public void scoreStudentParticipation(NusnetId nusnetId, int score) {
+    public void scoreStudentParticipation(NusnetId nusnetId, double score) {
         requireAllNonNull(nusnetId);
 
         boolean isFound = false;
@@ -108,7 +108,7 @@ public class StudentRecordListManager implements StudentRecordList {
      * Updates participation score of all students in the list of {@code nusnetIds} with {@code attendanceType}.
      */
     @Override
-    public void scoreAllParticipation(List<NusnetId> nusnetIds, int score) {
+    public void scoreAllParticipation(List<NusnetId> nusnetIds, double score) {
         for (NusnetId nusnetId : nusnetIds) {
             scoreStudentParticipation(nusnetId, score);
         }
